drop http request body early
Summary:
Minimal example of the problem: https://fb.workplace.com/groups/learningrust/permalink/27915999731355277/

I am doing this because we can OOM processing a 10G push:
https://pxl.cl/6C0RP

In memory profiles we see we hold the whole request body for the whole time of the request and response processing: https://fburl.com/scuba/heap_profiles/on_demand/3jz4nk9i

It's 21.7% memory of the tw task!

We don't fully stream everything. I think it's because gitoxide API doesn't support this. We however parse the body into Rust data structures, and one we finished reading it, we can discard it, rather than hold it for the lifetime of `body_bytes` (since that owns the underlying [u8]).

It generaly doesn't seem like the best pattern to propagate (by borrowing) references across structures and funcrtion calls. It's better to propagate wrapper types owning the underlying data via moving ownership. This is what we're doing by propagating `Bytes` now.

It also simplifies the signatures of functions.

Reviewed By: mitrandir77

Differential Revision: D69798635

fbshipit-source-id: 6d95ce7fbb239670e018e0eaf09d0badb716ac2a