Fix client-relative shallow clone/fetch in Mononoke Git server
Summary:
When users specify the `depth` flag when running `git clone` or `git pull`, the frontier of heads known by the server gets counted as depth=1. The second layer has depth=2 and so on.

However, when users specify `deepen` flag when running `git clone` or `git pull`, the shallow boundary at the client gets counted as depth=0. The second layer immediately below the shallow boundary has depth=1 and so on.

The bug between `depth` and `deepen` existed because of this difference. This diff fixes that.

Reviewed By: mitrandir77

Differential Revision: D69779644

fbshipit-source-id: 2ce1d4b43b3d90bd545f2ed9935f9b512717d34f