Integration test highlighting missing support for client-relative shallow clone/fetch
Summary:
The `deepen` flag can be used during `clone` or `pull` which behaves similar to the `depth` flag but with one major difference. The `depth` flag conveys the depth of the commit graph from the point of view of the server heads, whereas the `deepen` flag conveys the depth of the commit graph from the point of view of the client side shallow boundary.

This integration test highlights Mononoke Git server's shortcoming in producing the correct response as compared to vanilla Git when the `deepen` flag is employed.

Differential Revision: D69779356

fbshipit-source-id: 3e9bcb110cf996b270259ea8de83508700be99f8