Integration test highlighting lack of shallow-since support in Mononoke Git server
Summary:
`shallow-since` is a feature that allows users to clone repos in Git where the history of the repo is limited to commits that were created after a specific time stamp.

This is a relatively new feature and rarely implemented [properly in Git servers out there](https://stackoverflow.com/questions/73464420/git-fetch-shallow-clone-doesnt-pick-up-a-matching-commit). Additionally, this feature completely breaks if the repo has commits with non-monotonically increasing commit timestamps.

However, since its part of the Git protocol we have to support it  in some way. This diff adds an integration test that highlights our current lack of support for this feature.

Differential Revision: D69793472

fbshipit-source-id: 829ee8645e289ac8ffa5b84efa4b8e4d6e3f280c