add docs for CASC and SCM on RE
Summary:
add docs for CASC and SCM on RE

The plan is to write the documentation from the codehub UI (where the markdown preview is available), and then link the code source doc files to the wiki directly. All pictures will be uploaded and come from the pixelcloud links.

As a result, we will have **wiki in code** for these two projects

**Examples how this work:**

the wiki page: https://www.internalfb.com/wiki/Remote_Execution/engineering/compute/action_history_tracking backed by the doc in the code: https://fburl.com/code/zy9ccd8u

This approach has some clear advantages over the native wiki interface and also over the gdocs and syncing them to the wiki, because all the documentation will be available in both the code and the wiki.

Reviewed By: lmvasquezg

Differential Revision: D69920666

fbshipit-source-id: 4ae8bbdb7d6fe526bccd34f47ce11f2003ec3f0c