metaedit: allow individual commit message field updates
Summary:
Make "metaedit" (along with "commit" and "amend" ) support updating individual sections of the commit message. This makes it easy for automation to change the "Test Plan" section, for example, leaving all other sections untouched.

I went back and forth, but for now extra newlines must be specified by the user (e.g. if they want a blank line between "Section One: foo" and "Section Two: bar", they must add a newline at the end of their specification of "Section One"). This keeps things simple and makes it possible to do the newlines however you like, at the cost of it being less convenient for a user to use on the CLI (which is not the intended use case).

Reviewed By: quark-zju

Differential Revision: D69946882

fbshipit-source-id: d60bb3b87c3991f2eaaa70dc2e8c0dab8b1fd9c8