Prevent very large pushes to Mononoke Git server
Summary:
In Metagit, we limit the size of the push to be 3 GB. GitHub has a limit of 2GB by default. Currently, in Mononoke there is no limit at all so the user is free to push as much data as they want and overwhelm our server.

This diff adds the needed checks so if the packfile in the push is beyond the configurable limit of 7GB, the push gets rejected with appropriate error message.

Reviewed By: singhsrb, mzr

Differential Revision: D70092009

fbshipit-source-id: 9cb343d5400f32f24b584e2dcf3db49d3c074571