http-client: add simple request limiter
Summary: Add a "claimer" object that lets you claim a request spot. If there are spots available you get a "claim" which is intended to be stored with the request somehow. No matter how the request finishes, as long as the claim object is dropped then the request slot will be freed up.

Reviewed By: quark-zju

Differential Revision: D69927690

fbshipit-source-id: 1269c2060afef89a8582ee536429df03b940ce6c