checkout: fix progress bar
Summary:
I think this bar wasn't showing up because it was stuck in pending limbo.

The issue is that this code starts a bar, then updates the bar from different threads as it fans out work. Previously the thread-local "active" bar (ActiveProgressBar) didn't let you get the underlying ProgressBar out, so you couldn't share it with threads and had to jump through hoops (and this code wasn't doing the hoops right). This fan-out is relatively common, so let's just allow access to the underlying bar. The "danger" is if the bar is set as active again in another thread. Nothing bad would happen per se, but it wouldn't make sense.

Reviewed By: zzl0

Differential Revision: D69927700

fbshipit-source-id: 36bfec687ce537be554c63a2ff025e4cd6a429a1