progress: ensure space between topic and elapsed time
Summary:
Previously in the structured renderer the length of the elapsed time (e.g. "12s") would be counted against the topic length. This would end up truncating topic text even when topic length was below MAX_TOPIC_LENGTH, and fail to leave space between the topic and elapsed time.

Fix by budgeting for the elapsed time's length in the computed topic length, and prefixing the elapsed time with some spaces to ensure separation between the topic and elapsed time.

Reviewed By: quark-zju

Differential Revision: D69927689

fbshipit-source-id: 0fb489b3976be76a41efc13642130d4659c4de18