progress_model: add new "shared" flag for bars
Summary:
This flag is intended to tweak behavior for progress bars that are shared as the "active" bar across multiple threads.

By default, when a bar is popped (i.e. unset) as the active bar, the bar is also finished (i.e. state becomes Complete). This affects how the bar is rendered. If you want to share a bar across threads as the active bar in each thread, you don't want the bar to complete when the first thread finishes.

Reviewed By: zzl0

Differential Revision: D69927703

fbshipit-source-id: 2d5f521b18d483ba650ed05d36893152d10ef75d