http-client: limit response body buffering
Summary:
Previously we used unbounded queues for body chunks coming in from curl. This was convenient for never stalling out the requests, but obviously inconvenient when you have finite RAM.

Address by adding a limit to the queue and adding backpressure to curl by "pausing" the transfer when the queue is full. We unpause the transfer in our loop around the curl Multi object. Note that normally the unpausing happens in the curl "progress" callback, but the Rust Easy2 object we use doesn't make that possible/easy.

The new behavior defaults to "on", but you can disable by setting "http.limit-response-buffering=false".

Reviewed By: quark-zju

Differential Revision: D69927696

fbshipit-source-id: 2613936998139fde8188b7ecb66496efbe94220c