Implement read command to parse FAM output file
Summary:
## This Diff

Define function to deserialize the json string from a file into a vector of `Event` objects. The function handles the case when there is incomplete json string at the end.

## Context
`eden fam read -p /path/to/output/file` is a command to parse the output file of FAM into a summary report.

Reviewed By: genevievehelsel

Differential Revision: D69951073

fbshipit-source-id: 5b78fd88c031e0d7381c2b5b8db63606cb662e09