Add an option to specify the max number of events to display
Summary:
## This Diff
Add an option to specify the max number of events to display. This is helpful when there are a huge number of distinguished PIDs from the event objects.

## Context
`eden fam read -p /path/to/output/file` is a command to parse the output file of FAM into a summary report.

Reviewed By: genevievehelsel

Differential Revision: D69951072

fbshipit-source-id: 8784ba8209629dbfaaa3c7b9356615cfea027b98