Expose CASd only client in rust
Summary:
Instead of instantiating the REClient directly inside the rust bindings we now pass in a factory which can be used to produce a IREClient. This allows for use of the external CASd only RE client as a replacement for the existing external client.

To pass in a factory we:
- Have a rust struct that implements REClientFactoryTrait
- that trait has a method which returns a C++ factory (via ffi) implementing CppREClientFactory
- This is passed into the C++ code then used to create the client

As this is done by a struct, using a generic type on the REClientBuilder, so that we don't have to make all of the existing call sites any less ergonomic.

There are lots of lints here, but mostly in existing code. Will leave those for now

Sorry for the big diff, most is moving code

Differential Revision:
D69987427

Privacy Context Container: L1253485

fbshipit-source-id: 0146adcdce1d8ddfaf45750d3c06c16550cebdd1