Create and force indices for CLI queries
Summary:
## This stack

In S485666 we saw issues where the wrong index was being picked, which caused extremely long query times. D68768801 mitigated the issue.

As a followup I am auditing every other query. This stack addresses what I found.

## This diff

Some of these queries are expensive on a large result set. That affects CLI performance, and in SEV situations we've had to skip the CLI and go to the DB.

Reviewed By: lmvasquezg

Differential Revision: D70157040

fbshipit-source-id: f3fdedbfe019c65971cf32cf87eded495bf5e82b