Handle race conditions with the daemon starting and health-report runs
Summary:
If the user did a eden restart when periodic runs for health-report is executing, then users will receive a false notification for EdenFS being unhealthy.  Also, we need to make sure this won't run at login, meaning it wont race with eden start --if-necessary.

Solution:
Add a check for Eden starting to health-report and try to wait until EdenFS has started or the call timedout.

Reviewed By: genevievehelsel

Differential Revision: D69555032

fbshipit-source-id: 3556cb9b7cb4b8c99d9bc7fd3de8abd3b36e8de2