add a executor for checkoutRevision endpoint
Summary:
# Context
We’ve encountered issues in the past of certain functions having wide execution fan-out, meaning one Thrift call could tie up all of the Thrift CPU threads to serve the request. This is not ideal, as this results in other clients receiving `QueueTimeout`s due to waiting too long to get Thrift requests picked up for execution. We’ve combated this in the past by limiting some Thrift entry points to serial execution. This has become a game of whack-a-mole with new instances of `QueueTimeout`s, requiring copying this logic to different functions that seem to have similar fan-out, and is error prone since its easy to forget to set up a SerialExecutor while refactoring pre-existing functions.

There are some mission critical entry points that are known to be more heavyweight and should not be executed serially. There are a few options to be explored for these cases, including adding thread pools for particular heavyweight entry points, or implementing a Executor which can use smarter internal load balancing, such as a token bucket backed solution. This diff explores the first option.

# This Diff
This diff adds a `UnboundedQueueExecutor` that can be used to drive `checkout()`. Without this, and while using serial execution for Thrift requests across the board, checkout becomes too slow. This allows us to use serial execution for everything while keeping `checkoutRevision` performant. We might need to introduce similar executors for other endpoints, TBD.

# Technical Details
The use of a `UnboundedQueueExecutor` is explicit here. I plan to go through and remove uses of bounded executors in EdenFS. It is not recommended to use bounded executors with async code [1]

# Facebook
[1] - https://fb.workplace.com/groups/474291069286180/posts/5360947373953834/?comment_id=5361092820605956

[Threadpool Project Plan](https://docs.google.com/document/d/1yvYFYk8UT3hW9M7Qw-NSIpvqYFUxjHL8FhM81cYQGeQ/edit?tab=t.0)

Reviewed By: jdelliot

Differential Revision: D70135765

fbshipit-source-id: 66634542dbc3c69e8bb2c75ab071313e0ab37f80