Integration test highlighting lack of support for shallow-exclude in Mononoke Git
Summary:
I planned on creating this integration test to highlight that shallow-exclude is a shallow variant that is not yet supported in Mononoke, but during the course of its creation I realized that even vanilla git's support for shallow exclude is broken. I have linked a paste in the test that shows how git's actual behavior varies from what the documentation indicates.

Due to broken support in vanilla git, the integration test will compare the output of direct git command on the origin repo (`git rev-list --all --not branch_name`) instead of shallow-exclude clone using vanilla Git.

Differential Revision: D70256943

fbshipit-source-id: 7291f27d85c600dc33cd9ceca1d3bf55de327e19