Integration test highlighting vanilla Git behavior for commits with non-monotonic committer dates
Summary:
When performing a time-based shallow clone (or fetch), Git does a best effort clone of the commits if the repo contains commits with their committer dates not in monotonic ordering (i.e. earlier commits do not have a strictly lesser committer date than later commits).

This integration test highlights this behavior and serves as a proof that time-based shallow clones work only for the commits with monontonic committer dates. This would be useful for Mononoke Git cause we can assume that the graph only consists of such commits.

Differential Revision: D70242468

fbshipit-source-id: 6bf1a426fea13e35d0422cbd0804c987f687f4ce