Extend eden start to show notification when EdenFS ready
Summary:
Problem Statement
Laptop users frequently experience interruptions due to shutdowns and reboots, which can lead to unexpected issues with EdenFS. This can cause frustration when Eden is not ready to use after a restart.
Solution
To address this issue, we propose implementing a notification system that alerts users when EdenFS is ready to use after a restart or shutdown. This feature will be initially implemented on Windows systems.

Reviewed By: genevievehelsel

Differential Revision: D69509853

fbshipit-source-id: 3a8804b2c46cf404af5e9dd27ebb02f046588b89