Add probability distribution notebook
Summary:
# Context

The goal is to productionize the new **Performance** top-line metric for Source Control called User's Slowest Command (USC).

Details and principles on the top-line metric design and development are in this notebook - N6350927.

# Sequencing

We have to present the first version of this metric on **2025-Mar-5 OKR review**.

For that milestone we aim to build a dashboard, which will likely need.
1. Turn the notebook into a Python library so that it can be used both in notebooks and data pipelines.
1. Dedicated data pipeline as the `dev_command_timers` query is data and compute intensive.
1. Dashboard containing the top-line.
1. Sliced views by OS and which commands are contributing to the top-line.

After the dashboard.
1. Integrate the metric into Deltoid so that it shows up in DevInfra experiments and gradual rollouts.
1. Integrate in SLICK

# This diff

Add probability distribution notebook

Reviewed By: zzl0

Differential Revision: D70421910

fbshipit-source-id: 45f599b3e46baf8f8cc8c538f5d927efea5c1195