MercurialImporter Warnings and Future Proof until py3.16
Summary:
<frozen importlib._bootstrap>:1073: ImportWarning: MercurialImporter.find_spec() not found; falling back to find_module()

When you use `PYTHONWARNINGS=once` like I do, you see these kinds of messages.   Its not uncommon to see this message several times when using `arc pull`

But this is a warning to us that techdebt comes due eventually.    This deprecation takes places in python 3.12.  So it has to be fixed NOW.

The replacement of `load_module` sets us up for 3.15+ support

Reviewed By: itamaro

Differential Revision: D70416251

fbshipit-source-id: 60321abc6724e9c7de72a52e526e31a255bbc426