modern sync: checkpoint latest
Summary:
This ended up being waaay more complex than I thought when I realized we need to checkpoint __after__  changeset has been synced instead of after the changeset has been enqueued. Otherwise if we resume from a failed changeset it'll harder to backtrack.

So I ended up doing what we know how to do best in modern sync: add channels. This time we have a task that simply reads positions from a mpsc channel and then sets the counter to that. We send to this channel after uploading a batch of changesets and emptying `pending_checkpoints` which we have accumulated while forming the batch.

Reviewed By: andreacampi

Differential Revision: D70534924

fbshipit-source-id: 00f9e09576fbf4922190f3bff3fe7f850f004833