Add a notebook used to generate numbers for the OKR review
Summary:
# Context

The goal is to productionize the new **Performance** top-line metric for Source Control called User's Slowest Command (USC).

Details and principles on the top-line metric design and development are in this notebook - N6350927.

# Sequencing

We have to present the first version of this metric on **2025-Mar-5 OKR review**.

For that milestone we aim to build a dashboard, which will likely need.
1. Turn the notebook into a Python library so that it can be used both in notebooks and data pipelines.
1. Dedicated data pipeline as the `dev_command_timers` query is data and compute intensive.
1. Dashboard containing the top-line.
1. Sliced views by OS and which commands are contributing to the top-line.

After the dashboard.
1. Integrate the metric into Deltoid so that it shows up in DevInfra experiments and gradual rollouts.
1. Integrate in SLICK

# This diff

Add a notebook used to generate numbers for the OKR review

Reviewed By: MichaelCuevas

Differential Revision: D70588126

fbshipit-source-id: c10dd04383aa089c32d5d050d092d94747268e62