tests: move edenfs-related parts of debugruntest to an extension
Summary:
Similar to what we do for the Mononoke extension. We'll add a bit more of EdenFS logic to debugruntest, so let's set up a nice precedent before continuing.

The rest of the logic in `tests/run-tests.py` and `tests/edenfs.py` cannot yet be moved to this extension. See comments in `tests/edenfs.py` as to why.

Reviewed By: muirdm

Differential Revision: D70488446

fbshipit-source-id: 7f968446e83c6507a983a7127f04acc75be4a9b6