tests: get rid of redundant `exe=True` in hghave's checks for EdenFS
Summary:
Using `, exe=True` in a `check` in hghave adds that command to the commands table. This is useful if using that hghave condition as part of `#require`s.
Since now some tests run with EdenFS by default without having those conditions, we added an additional `t.requireexe` for the EdenFS CLI back in D56660786

The additional `, exe=True` is not really necessary, since it can be added anyways through `t.requireexe` simplifying things a bit.

Reviewed By: quark-zju

Differential Revision: D70488447

fbshipit-source-id: 67516d37687c8cb38ff258fc9b9508981687c545