website: make website compatible with newer JS versions
Summary:
The job for building the website internally is failing with

```
  [cause]: [Error [ValidationError]: "remarkPlugins[0]" does not look like a valid MDX plugin config. A plugin config entry should be one of:  - A tuple, like `[require("rehype-katex"), { strict: false }]`, or
  - A simple module, like `require("remark-math")`] {
```

due to an internal JS upgrade. The `remark-github` script [hasn't been upgraded in a while](https://github.com/remarkjs/remark-github), so it makes sense that it fails with a new JS version.

Let's just use the default function for walking around this.

Reviewed By: muirdm

Differential Revision: D70662751

fbshipit-source-id: 0ecdc0eb7938fb6db4d822032d67ac654ebbbbb8