remove ability to use a bounded executor for the FsChannel
Summary:
# Project Context
EdenFS has a non-trivial threading model. It has a series of executors, some which have very specific purposes and some which handle generic work. Specific units of work (such as, serving a Thrift call) can jump around from one executor to another. A handful of Thrift entrypoints, like the entrypoints that handle prefetch and Watchman queries, have wide fanout, and can monopolize EdenFS' resources, leading to starvation and eventual Queue Timeouts for other Thrift methods.

Furthermore, there exist foot-guns in EdenFS that have caused repeated SEVs (S412223, S399431, S401889) and have caused user-visible issues like deadlocks and exceptions (D50199539, D41390586).  These issues are easy to accidentally introduce due to the lack of guard rails in place and are difficult to root cause. Much of our threading rules/logic is mandated by in-code comments or learnt prior experience, which is not scalable and makes it hard to onboard and confidently land changes. This project's goal is to de-risk changes that touch these parts of the codebase and to make it easier to understand and optimize EdenFS' resource usage.

# Subtask Context
Most of our thread pools are managed by Folly executors. There are bounded and unbounded executors. We’ve encountered SEVs in the past where trying to enqueue work onto a full, bounded executor can cause a deadlock (S412223). To avoid these suites of problems, we can ensure all of our executors are unbounded and add backpressure manually at EdenFS' two entry points (Thrift and FSChannel) by limiting the amount of work that enters EdenFS using configurable semaphores.

# This Diff
Using bounded queues is deadlock prone and not recommended while using Folly async frameworks. This diff deletes the ability to use a bounded queue in the FSChannel.

# Technical Details
In response to S412223, we unbounded the FSChannel thread pool. The bounded threadpool caused a deadlock (see D56553375 for more details). tl;dr we were deadlocked because the FsChannelThreads were trying to enqueue onto a full task queue, but that task queue could only be emptied by the FsChannelThreads themselves.

There is a lot of discussion that can be found on Workplace about the use of bounded queues vs unbounded queues, but the suggestion across the board is to avoid bounded queue executors when using folly async frameworks, as bounded queues are prone to the deadlock outlined above. [1] [2]

# Resources

Reviewed By: MichaelCuevas

Differential Revision: D70664180

fbshipit-source-id: 8234d236cc34831e43d9885dc892d7a06d6f27f0