add inode invalidation
Summary:
# Context
Currently the periodic Garbage Collection (GC) only works for windows. We need to extend this periodic process to macOS platform to prevent macOS crashes and eden instability when a crawler loads millions of inodes.

# This Diff
The last step of GC is cache invalidation and dereferencing the inodes. Instead of modifying `invalidateChannelEntryCache`, I create `nfsInvalidateCacheEntryForGC` only for NFS. Because `invalidateChannelEntryCache` is used in other places in `TreeInode.cpp` and I don't want to interfere with other code in production. The GC on NFS only run from `eden debug gc-inode` command for the users who activate `enable-garbage-collection` config

Reviewed By: genevievehelsel

Differential Revision: D70260880

fbshipit-source-id: a32a24c23a0409e406776ad272cdf5f320f01022