make fs refcount a flag on NFS
Summary: On macOS, the fs refcount only increased by `NfsDispatcherImpl.cpp` and we don't have any fs refcount decrement on NFS. The only time that we call `decFsRefcount` on NFS is when we want to clear the fs refcount. Therefore, it is better to make the fs refcount a flag on NFS same as what we have on windows. See summary of D24764567 for more details oin windows

Reviewed By: genevievehelsel

Differential Revision: D70585532

fbshipit-source-id: a4b25820c1ae7e0ef668adb6666cbe8585e53203