add onSuccess to NFS invalidation
Summary:
# Context
Currently the periodic Garbage Collection (GC) only works for windows. We need to extend this periodic process to macOS platform to prevent macOS crashes and eden instability when a crawler loads millions of inodes.

# This Diff
NFS invalidation is a hack for more details see the summary of D35435764
This diff adds a lambdafunction to the NFS invalidation, and the NFS invalidation run this function only it is successful. I will use this function in the next diff.

Reviewed By: genevievehelsel

Differential Revision: D69537035

fbshipit-source-id: 0723587661f7ebd2b9a61208ea78b7a01e6493f6