read the atime from the InodeMetadata table
Summary:
# Context
Currently the periodic Garbage Collection (GC) only works for windows. We need to extend this periodic process to macOS platform to prevent macOS crashes and eden instability when a crawler loads millions of inodes.

# This Diff
For non-windows platforms we have InodeMetadata table, and we can read the atime from the InodeMetadata table. It helps us to skip inode invalidation if they are accessed recently.

Reviewed By: lXXXw

Differential Revision: D69265056

fbshipit-source-id: 1b20c0b9db3c862b88e2bf97392a765259c523a3