Only upload if the batch is not empty
Summary:
## This stack

Log more details on the sync process to help with debugging.

## This diff

In the typical case we will receive messages in the ideal sequence:

* `FileMessage::FileNode`
* `TreeMessage::TreesDone`
* build a batch and call `flush_batch`

This is the happy path, and everything works.

Occasionally though this would work out differently:

* `FileMessage::FileNode`
* build a batch and call `flush_batch`
* `TreeMessage::TreesDone`
* build a batch and call `flush_batch`

I am not clear yet why, but this sequence would result in a hang. A similar sequence is possible for trees.

This diff ensure we only try to upload non-empty batches, which prevents the issue.

Reviewed By: lmvasquezg

Differential Revision: D70849268

fbshipit-source-id: 91b64c10a7b6a0668d65919fe3670b9baa3b214e