Back out "Revert D70534924"
Summary:
On Friday I thought D70534924 might be causing lockups, but this turned out to be a red herring.

Now we have a proper root cause and a fix-reply apply the change. We do still need it to speed up the recovery on restart and on errors.

In the meantime we've had enough changes that I had to resolve a number of conflicts. I compared each hunk against D70534924 to make sure the change is functionally the same, regardless of moving to a different function and so.

Reviewed By: lmvasquezg

Differential Revision: D70887211

fbshipit-source-id: bf5dec483c9d4304c3ceadc99c59376292f0c13f