Make path elements with "\n" or "\x01" representable in the data model
Summary:
These characters conflict with Sapling's data model, but not with the world.

As Mononoke is a multi-front-end Source Control backend, such paths should be representable.

In particular, the `ssziparchive` Git repository contains a filename containing a "\n", which is how this came to our intention.

Of course, we still don't want Sapling users to create such commits, which is why:
* Sapling itself doesn't let users generate such commits. See [code](https://fburl.com/code/hf7kytie) and [repro](P1748098007)
* We added a default hook with no bypass string or PUSHVAR which will reject such commits on the Mononoke side. See D70697444
* We are explicitely failing derivation of HgManifests when such bytes are present to avoid generating invalid Hg changesets. See D70712242

The aim of this diff is only to make it representable so that we can import commits coming from Open-Source Git repositories that may hit this corner case.

Reviewed By: RajivTS

Differential Revision: D70712854

fbshipit-source-id: 4795044fc9a55ae9bae05294e5d729d3d97cb534