Include retries for fetching existing bonsai git mappings during Mononoke Git push
Summary:
The first step of the gitimport process of Mononoke Git push involves fetching the existing bonsai git mappings based on the input set of commits. For the commits where we already have a mapping, we skip the import process because its done already.

For large pushes (e.g. `chromium/src`), this involves fetching thousands of preexisting mappings. In such cases, the query for fetching the mappings might time out or get rejected. Instead of failing the push in that case, let's try retrying the fetch a few times before giving up.

Differential Revision: D70890252

fbshipit-source-id: 4b0b4918df5059cfa8461d35dc0d7d832714ce3c