add loaded inodes number to fetch heavy events
Summary:
When we record fetch_heavy events in scuba, we record the fetch_count. However fetch_count is not the number of loaded inodes. To debug issue, we need to know the number of loaded inodes in these event to find out the cmd line which is responsible for loading many inodes.
This diff adds loaded inodes number to the edenfs_events scuba table for fetch_heavy events.

Reviewed By: jdelliot

Differential Revision: D70679540

fbshipit-source-id: d816f5f999ecc8fde2693bbd2e020cbc01f2dd16