modern sync: enqueue post processing
Summary:
{F1975959735}

Several things going on here, I'll try to describe them as best as I can.

### Bookmark moving

If we move master from `A -> B` and then from `B -> C` instead of waiting for B to sync, moving the bookmark, and then waiting for C and moving the bookmark, we'll simply enqueue `[A,B,C]` and `[A->B,B->C]` which, if they're processed in the same batch will translate into `[A->C]`

### Checkpointing

Given that there's a chance we checkpoint and entry thas has been completely enqueued I'm introducing a new counter `CURRENT_ENTRY` that indicates which entry does the `BATCH_CHECKPOINT` value correspond to. E.g
Entry 1: `[A, B, C, D]` Entry 2: `[E, F, G]`

If while finish syncing B and checkpoint only `BATCH_CHECKPOINT=1` there's a chance we take that as the index to start Entry 2, skipping commit F. While if we checkpoint  `CURRENT_ENTRY=1` + `BATCH_CHECKPOINT=1` if we're enqueueing Entry 2 we can know that's not a valid checkpoint and simply start from zero.

Also if we enqueue [ Entry:1 - Position:1, Entry:2 - Position:1] we're eventually just set Entry:2 - Position:1 if there's processed in the same batch

This counter is also reset to zero after finishing each bookmark move.

### Completion

In cases where we want to terminate the sync once all bookmarks  are synced we can't simply terminate the main executor because most likely all element haven't finished syncing, hence we add a last message to the queue  that indicates all elements have been sent out and it's safe to terminate the executor.

Reviewed By: andreacampi

Differential Revision: D70941608

fbshipit-source-id: ec857d58915c4df492b70d0920bd34b21912bcc4