Pull in buck2 ErrorHandlingStrategy and related
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Pull in buck2 ErrorHandlingStrategy - a mechanism for defining how we handle errors returned from Thrift. This is needed before we can pull in the buck2 with_eden (we will call this with_client) - a richer mechanism for lazily creating connections and handling retries and throttling.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: genevievehelsel

Differential Revision: D70917966

fbshipit-source-id: cc429f72c55e5e3fdd7be786d3b211ad2fef907c