Make changesets public and derive stuff
Summary:
This diff is designed to help with modern sync (currently the only use case for this endpoint). After syncing changesets, we will move the bookmark, and that will cause WBC to perform both of these operations. However, that is not deterministic, and it's likely to be slow–possibly even timeout.

As an added benefit, doing this progressively means we'll be able to `sl clone` then `sl goto` even before the first bookmark move.

Reviewed By: markbt

Differential Revision: D70993804

fbshipit-source-id: d29311da9fb5a0f05825b39e4542a7d4bff87330