universal SQL impl. facet impl
Summary:
I'm simplifying the bundle-uri functionality.
Instead of a generic library that does too much with some anticipated use-cases, let's just have one that's very simple and does the thing we need now.

I'm pulling up the sql.rs outside of `facebook` directory. It has nothing that's really Meta-specific. I'm also simplifying it quite a bit. There will be no async thread keeping the state fresh in memory, to be more in line with our design patterns elsewhere. It will also be purely per-repo, because I'm making the BundleURI structure a facet - we don't have an immediate need to run it outside of MononokeApp.

Reviewed By: RajivTS

Differential Revision: D70962684

fbshipit-source-id: a2a48af96d64226ecc08b30ba7cd4dd0dac491f1