add rate limiting ability to FuseChannel
Summary:
# Project Context
EdenFS has a non-trivial threading model. It has a series of executors, some which have very specific purposes and some which handle generic work. Specific units of work (such as, serving a Thrift call) can jump around from one executor to another. A handful of Thrift entrypoints, like the entrypoints that handle prefetch and Watchman queries, have wide fanout, and can monopolize EdenFS' resources, leading to starvation and eventual Queue Timeouts for other Thrift methods.

Furthermore, there exist foot-guns in EdenFS that have caused repeated SEVs (S412223, S399431, S401889) and have caused user-visible issues like deadlocks and exceptions (D50199539, D41390586).  These issues are easy to accidentally introduce due to the lack of guard rails in place and are difficult to root cause. Much of our threading rules/logic is mandated by in-code comments or learnt prior experience, which is not scalable and makes it hard to onboard and confidently land changes. This project's goal is to de-risk changes that touch these parts of the codebase and to make it easier to understand and optimize EdenFS' resource usage.

# Subtask Context
Most of our thread pools are managed by Folly executors. There are bounded and unbounded executors. We’ve encountered SEVs in the past where trying to enqueue work onto a full, bounded executor can cause a deadlock (S412223). To avoid these suites of problems, we can ensure all of our executors are unbounded and add backpressure manually at EdenFS' two entry points (Thrift and FSChannel) by limiting the amount of work that enters EdenFS using configurable semaphores.

# This Diff
This diff adds a `RequestPermitVendor` to `FSChannel` and teaches `FuseChannel` to acquire a request permit before executing a FUSE request. The `RequestPermit` is held for the duration of the execution and is automatically released after the future has been fulfilled.

This diff also changes the default value of `fschannel:max-inflight-requests` to `0` - making the default behavior to mimic what we have today in production, no rate limiting by default.

# Resources

Reviewed By: jdelliot

Differential Revision: D70812137

fbshipit-source-id: 6c469c0a2ae3fdc2fddd0535e323d8cd41f5aaec