Move --debug and --verbose to the server
Summary:
Problem: ISL in vscode wouldn't load, as it was transitively importing WebSocket.

This was caused by using the debug & verbose atoms in Operation.tsx, which apparently is needed by the vscode extension. But if that loads platform, it would load WebSocket.

So we need to remove any client-specific stuff from Operation.tsx. That means we can't get the --debug and --verbose flags from jotai. Let's instead sync these to the server ctx from the client.

This isn't necessarily ideal, but it does fix the problem for now. I did this instead of finding all teh places that call `getRunnableOperation` and augmenting them with more arguments, which seemed silly.

Differential Revision: D71062977

fbshipit-source-id: 236d979d55f5020cc8bfbe46d48f170d7df90290