Migrate to with_eden and with_streaming_eden pattern
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

buck2 had a nice way of connecting to the thrift service and invoking endpoints there. Lets copy that. Here we move to the pattern without all of the goodies - retries, reconnects, etc.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: genevievehelsel

Differential Revision: D70941230

fbshipit-source-id: 5a23aa67b891344720e0b34ddf0c1595a99af1a7