pager: fix default pager.stderr config
Summary:
This only defaulted to `true` in Python, but pager handling has been moved to Rust. Move the default into Rust.

This fixes apparent interleaving of output as stdout goes to the pager and stderr goes directly to the terminal. Note that this does not affect streampager which always receives stderr.

Reviewed By: zzl0

Differential Revision: D71087524

fbshipit-source-id: bbc85abce23f2d11a9ffcee517500fd6f2d7f19d