modern sync: fix duped upload
Summary:
So in the latest diff we made the sync resume from latests `master`. However, we can't guarantee latest `master` matches the end of the previous entry due to the lag while all the elements get sent, causing that segments were being calculated on a stale master and being enqueue twices, which generally is not a huge deal unless they're sent in the same batch, cause if we upload [A,B,B] and receive [A,B] as "uploaded", we'll fail.

I *could*  simple dedupe the changeset request before sending it, but we're wasting time re-reading each duped commit, so I opted for simply keeping track of the latest commit of the previous entry and calculating segments using it given that `ancestors_difference_segment_slices` does not include `commons`, so no duped commits.

So in the new mechanism, if we have a BUL like:
```
1. None->A (master)
2. None->B
3. None->C
4. None->D
```

And resume the sync, before we's have segments `[B]` for 2 and `[B,C]` for 3 and `[B,C,D]` for 4.
 Now, we'll only have `[B]` for 2 and `[C]` for 3 and `[D]` for 4

Reviewed By: andreacampi

Differential Revision: D71048488

fbshipit-source-id: 2e23476d324faa6fb1faf5a426a655dd60694cd0