Improve error reporting for glob_files
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

In a previous diff, we increased error reporting from a cleint (meerkat), but sadly this only pushed the issue down in the stack. Let's fix that by including the params and error when glob_files fails.

Later, we should move to more structured errors (not anyhow!) so that we can be more explicit in what failed.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Differential Revision: D71128650

fbshipit-source-id: f0ac8fefc53808646ece907ad474e34a7327a5f7