rust: fix clippy warn borrow_interior_mutable_const
Summary:
This seems a bug in our code not just a style warning.

According to the clippy doc [1], the `const` are copied everywhere they are referenced, i.e.,
every time you refer to the const a fresh instance of the Cell or Mutex or AtomicXxxx will be
created, which defeats the whole purpose of using these types in the first place.

For example if one calls `set_batch_size()`, the `BATCH_SIZE.load(Ordering::Acquire)` will
still get the old value. [2]

[1] https://rust-lang.github.io/rust-clippy/master/index.html#borrow_interior_mutable_const
[2] rust playground test: https://play.rust-lang.org/?version=stable&mode=debug&edition=2024

Reviewed By: quark-zju, sggutier

Differential Revision: D71044268

fbshipit-source-id: 94619a174a0a160e05742ba830776d84adc5bbf6